update iOS example
