+++ Examples/CollectionView/SkeletonViewExampleCollectionview-Info.plist
