+++ Examples/iOS Example/Sources/Assets.xcassets/avatar.imageset/Contents.json
