+++ Examples/iOS Example/Sources/Assets.xcassets/avatar.imageset/avatar.png
